types: move all definitions out of the index
