+++ modules/types/src/convert.ts
@@ -2,14 +2,15 @@ import { AddressZero } from "ethers/constants";
 import { BigNumber, getAddress } from "ethers/utils";
 
 import { AppState, CoinTransfer } from "./app";
+import { AssetAmount, MultisigState, PaymentProfile } from "./channel";
 import {
   DepositParameters,
   LinkedTransferParameters,
+  LinkedTransferToRecipientParameters,
   SwapParameters,
   TransferParameters,
   WithdrawParameters,
 } from "./inputs";
-import { AssetAmount, MultisigState, PaymentProfile } from "./node";
 
 /////////////////////////////////////////////
 ///////// CONVERSION FNS
@@ -200,6 +201,13 @@ export function convertLinkedTransferParametersToAsset<To extends NumericTypeNam
   return convertAssetAmountWithId(to, obj);
 }
 
+export function convertLinkedTransferToRecipientParametersToAsset<To extends NumericTypeName>(
+  to: To,
+  obj: LinkedTransferToRecipientParameters<any>,
+): LinkedTransferToRecipientParameters<NumericTypes[To]> {
+  return convertAssetAmountWithId(to, obj);
+}
+
 export function convertWithdrawParametersToAsset<To extends NumericTypeName>(
   to: To,
   obj: WithdrawParameters<any>,
@@ -213,7 +221,10 @@ export function convertAppState<To extends NumericTypeName>(
 ): AppState<NumericTypes[To]> {
   return {
     ...obj,
-    transfers: [convertAmountField(to, obj.transfers[0]), convertAmountField(to, obj.transfers[1])],
+    coinTransfers: [
+      convertAmountField(to, obj.coinTransfers[0]),
+      convertAmountField(to, obj.coinTransfers[1]),
+    ],
   };
 }
 
@@ -223,6 +234,7 @@ export const convert = {
   Asset: convertAssetAmount,
   Deposit: convertDepositParametersToAsset,
   LinkedTransfer: convertLinkedTransferParametersToAsset,
+  LinkedTransferToRecipient: convertLinkedTransferToRecipientParametersToAsset,
   Multisig: convertMultisig,
   PaymentProfile: convertPaymentProfile,
   ResolveLinkedTransfer: convertAssetAmountWithId,
